# openapi.yaml
...
openapi: 3.0.0
info:
  x-skaffolder-id-project: 5dcb9895f1ef4518a5382d44
  title: Todo
  description: You can obtain JWT token from login API
  version: 0.1
servers:
  - url: http://localhost:3000/api
    description: Local server for testing
paths:
...
  /user/:
    post:
      x-skaffolder-id: 5dcb9896f1ef4518a5382d53
      x-skaffolder-name: create
      x-skaffolder-id-resource: 5dcb9895f1ef4518a5382d4e
      x-skaffolder-resource: User
      x-skaffolder-crudAction: create
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION create
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION create
      tags:
        - User
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                mail:
                  type: string
                name:
                  type: string
                password:
                  type: string
                roles:
                  type: string
                surname:
                  type: string
                username:
                  type: string
    get:
      x-skaffolder-id: 5dcb9896f1ef4518a5382d51
      x-skaffolder-name: list
      x-skaffolder-id-resource: 5dcb9895f1ef4518a5382d4e
      x-skaffolder-resource: User
      x-skaffolder-crudAction: list
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION list
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION list
      tags:
        - User
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
  /user/{id}:
    delete:
      x-skaffolder-id: 5dcb9896f1ef4518a5382d57
      x-skaffolder-name: delete
      x-skaffolder-id-resource: 5dcb9895f1ef4518a5382d4e
      x-skaffolder-resource: User
      x-skaffolder-crudAction: delete
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION delete
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION delete
      tags:
        - User
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      x-skaffolder-id: 5dcb9896f1ef4518a5382d54
      x-skaffolder-name: get
      x-skaffolder-id-resource: 5dcb9895f1ef4518a5382d4e
      x-skaffolder-resource: User
      x-skaffolder-crudAction: get
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION get
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION get
      tags:
        - User
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    post:
      x-skaffolder-id: 5dcb9896f1ef4518a5382d55
      x-skaffolder-name: update
      x-skaffolder-id-resource: 5dcb9895f1ef4518a5382d4e
      x-skaffolder-resource: User
      x-skaffolder-crudAction: update
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION update
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION update
      tags:
        - User
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
    
    
# JWT security
security:
  - bearerAuth: [] 
  
  
# Components
components:
...
  securitySchemes:
    bearerAuth:            # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT    # optional, arbitrary value for documentation purposes

  # Data Model
  schemas:
...
    User:
      x-skaffolder-id: 5dcb9895f1ef4518a5382d4e
      x-skaffolder-id-db: 5dcb9895f1ef4518a5382d45
      x-skaffolder-id-entity: 5dcb9895f1ef4518a5382d46
      x-skaffolder-url: /user
      properties:
        _id:
          type: string
          x-skaffolder-required: true
        mail:
          type: string
          x-skaffolder-id-attr: 5dcb9895f1ef4518a5382d49
          x-skaffolder-type: String
        name:
          type: string
          x-skaffolder-id-attr: 5dcb9895f1ef4518a5382d4a
          x-skaffolder-type: String
        password:
          type: string
          x-skaffolder-id-attr: 5dcb9895f1ef4518a5382d48
          x-skaffolder-type: String
          x-skaffolder-required: true
        roles:
          type: string
          x-skaffolder-id-attr: 5dcb9895f1ef4518a5382d4c
          x-skaffolder-type: String
        surname:
          type: string
          x-skaffolder-id-attr: 5dcb9895f1ef4518a5382d4b
          x-skaffolder-type: String
        username:
          type: string
          x-skaffolder-id-attr: 5dcb9895f1ef4518a5382d47
          x-skaffolder-type: String
          x-skaffolder-required: true
      x-skaffolder-relations: 
      required:
        - _id
        - password
        - username
      
    
    
  # Databases
...